chore(ruby): update ruby-model seed
