+++ seed/ruby-model/ruby-reserved-word-properties/lib/seed_ruby_reserved_word_properties_client.rb
@@ -1,3 +1,3 @@
 # frozen_string_literal: true
 
-require_relative "seed_ruby_reserved_word_properties_client/types/foo"
+require_relative "seed_ruby_reserved_word_properties_client/service/types/foo"
